added cd material prop
