+++ src/Common/Types.ts
@@ -1,7 +1,7 @@
 import React, { ReactNode, CSSProperties } from 'react'
 import { Placement } from 'tippy.js'
 import { ImageComment, ReleaseTag } from './ImageTags.Types'
-import { DockerConfigOverrideType, SortingOrder, TaskErrorObj } from '.'
+import { ACTION_STATE, DEPLOYMENT_WINDOW_TYPE, DockerConfigOverrideType, SortingOrder, TaskErrorObj } from '.'
 
 /**
  * Generic response type object with support for overriding the result type
@@ -174,7 +174,6 @@ export interface RadioGroupComposition {
     Radio?: React.FC<any>
 }
 
-
 export interface RadioGroupProps {
     children: ReactNode
     value: string
@@ -276,6 +275,12 @@ export enum MaterialDataSource {
     EXTERNAL = 'ext',
 }
 
+export interface DeploymentWindowArtifactMetadata {
+    id: number
+    name: string
+    type: DEPLOYMENT_WINDOW_TYPE
+}
+
 export interface CDMaterialType {
     index: number
     id: string
@@ -313,6 +318,7 @@ export interface CDMaterialType {
     createdTime?: string
     deployed?: boolean
     dataSource?: MaterialDataSource
+    deploymentWindowArtifactMetadata?: DeploymentWindowArtifactMetadata
 }
 
 export enum CDMaterialServiceEnum {
@@ -806,3 +812,13 @@ export type SortingParams<T = string> =
           sortBy: T
       }
     | { sortOrder?: never; sortBy?: never }
+
+export interface DeploymentWindowProfileMetaData {
+    name: string
+    userActionState: ACTION_STATE
+    type: string | DEPLOYMENT_WINDOW_TYPE
+    calculatedTimestamp: string
+    isActive: boolean
+    excludedUserEmails: string[]
+    warningMessage: string
+}
